+++ src/App.jsx
@@ -2,11 +2,11 @@ import { useState } from 'react'
 import { Layers } from 'three'
 import { Canvas } from '@react-three/fiber'
 import { Physics, Debug } from '@react-three/cannon'
-import { Sky, Environment, PerspectiveCamera, OrthographicCamera, OrbitControls } from '@react-three/drei'
+import { Sky, Environment, PerspectiveCamera, OrthographicCamera, OrbitControls, Stats } from '@react-three/drei'
 import { useStore, levelLayer } from './store'
-import { Ramp, Track, Vehicle, Goal } from './models'
-import { HUD, Minimap, Overlay } from './ui'
-import { Controls } from './controls'
+import { Ramp, Track, Vehicle, Goal, Train, Heightmap } from './models'
+import { Clock, Speed, Minimap, Intro, Help, Editor, LeaderBoard, Finished } from './ui'
+import { HideMouse, Keyboard } from './controls'
 
 const layers = new Layers()
 layers.enable(levelLayer)
@@ -18,9 +18,9 @@ function DebugScene({ children }) {
 
 export function App() {
   const [light, setLight] = useState()
-  const [shadows, dpr, camera, editor, map] = useStore((s) => [s.shadows, s.dpr, s.camera, s.editor, s.map])
+  const [shadows, dpr, camera, editor, map, finished, stats] = useStore((s) => [s.shadows, s.dpr, s.camera, s.editor, s.map, s.finished, s.stats])
   return (
-    <Overlay>
+    <Intro>
       <Canvas key={shadows + dpr} mode="concurrent" dpr={[1, dpr]} shadows={shadows} camera={{ position: [0, 5, 15], fov: 50 }}>
         <fog attach="fog" args={['white', 0, 500]} />
         <Sky sunPosition={[100, 10, 100]} scale={1000} />
@@ -46,18 +46,27 @@ export function App() {
               <PerspectiveCamera makeDefault={!editor && camera !== 'BIRD_EYE'} fov={75} rotation={[0, Math.PI, 0]} position={[0, 10, -20]} />
               <OrthographicCamera makeDefault={!editor && camera === 'BIRD_EYE'} position={[0, 100, 0]} rotation={[(-1 * Math.PI) / 2, 0, Math.PI]} zoom={15} />
             </Vehicle>
+            <Train />
             <Ramp args={[30, 6, 8]} position={[2, -1, 168.55]} rotation={[0, 0.49, Math.PI / 15]} />
-            <Track position={[0, 0, 0]} />
+            <Heightmap elementSize={0.5085} position={[327 - 66.5, -3.3, -473 + 213]} rotation={[-Math.PI / 2, 0, -Math.PI]} />
             <Goal start args={[0.001, 10, 18]} rotation={[0, 0.55, 0]} position={[-27, 1, 180]} />
             <Goal args={[0.001, 10, 18]} rotation={[0, -1.2, 0]} position={[-104, 1, -189]} />
           </DebugScene>
         </Physics>
+        <Track />
         <Environment files={'textures/dikhololo_night_1k.hdr'} />
         {map && <Minimap />}
         {editor && <OrbitControls />}
       </Canvas>
-      <HUD />
-      <Controls />
-    </Overlay>
+      <Clock />
+      {editor && <Editor />}
+      {finished && <Finished />}
+      <Help />
+      <Speed />
+      {stats && <Stats />}
+      <LeaderBoard />
+      <HideMouse />
+      <Keyboard />
+    </Intro>
   )
 }
