+++ src/controls/index.jsx
@@ -1,9 +1,2 @@
-import { Keyboard } from './Keyboard'
-import { HideMouse } from './HideMouse'
-
-export const Controls = () => (
-  <>
-    <HideMouse />
-    <Keyboard />
-  </>
-)
+export * from './Keyboard'
+export * from './HideMouse'
