+++ src/models/Goal.js
@@ -9,9 +9,7 @@ export function Goal({ start, args = [1, 1, 1], ...props }) {
       mutation.finish = 0
     } else {
       mutation.finish = Date.now()
-
       const time = !mutation.start && !mutation.finish ? 0 : mutation.finish ? mutation.finish - mutation.start : Date.now() - mutation.start
-
       set({ finished: time })
     }
   }
