+++ src/models/track/index.js
@@ -1,2 +1,3 @@
 export * from './Heightmap'
 export * from './Track'
+export * from './Train'
