+++ src/ui/LeaderBoard.jsx
@@ -5,29 +5,26 @@ import { getLeaderBoardData } from '../utils/data/leaderboard'
 export function LeaderBoard() {
   const [leaderBoard, setLeaderBoard] = useState([])
   const [set, leaderboard] = useStore((state) => [state.set, state.leaderboard])
-
-  const loadLeaderBoard = async () => {
-    const leaderboardData = await getLeaderBoardData(10)
-    setLeaderBoard(leaderboardData)
-  }
   useEffect(() => {
+    async function loadLeaderBoard() {
+      const leaderboardData = await getLeaderBoardData(10)
+      setLeaderBoard(leaderboardData)
+    }
     loadLeaderBoard()
   }, [])
   return (
-    <>
-      <div className="controls">
-        <div className={`popup ${leaderboard ? 'open' : ''}`}>
-          <button className="popup-close" onClick={() => set({ leaderboard: false })}>
-            L
-          </button>
-          <ul className="popup-content leaderboard-bottom">
-            {leaderBoard.map((item) => (
-              <Item {...item} key={item.id} />
-            ))}
-          </ul>
-        </div>
+    <div className="controls">
+      <div className={`popup ${leaderboard ? 'open' : ''}`}>
+        <button className="popup-close" onClick={() => set({ leaderboard: false })}>
+          L
+        </button>
+        <ul className="popup-content leaderboard-bottom">
+          {leaderBoard.map((item) => (
+            <Item {...item} key={item.id} />
+          ))}
+        </ul>
       </div>
-    </>
+    </div>
   )
 }
