+++ src/ui/index.js
@@ -1,3 +1,8 @@
-export * from './HUD'
+export * from './Clock'
+export * from './Speed'
+export * from './Editor'
+export * from './Finished'
+export * from './Help'
 export * from './Minimap'
-export * from './Overlay'
+export * from './Intro'
+export * from './LeaderBoard'
